change filenames to reflect the repo pattern more accurately + use entity when you need its fields (its a data representation of the holdings table)
